requirements.txt
fastapi
uvicorn
openai
python-dotenv
zipfile36
python-multipart
pandas 